Remove duplicate invocation of getKey() in File::getValueFromCache()
Calling `getKey()` before passing the key to `getFileName()` from `getValueFromCache()` is not only unnecessary, it actually breaks caching if the prefix is not empty.
I discovered this from a slightly different perspective: in my application I had to inherit from `Adapter/File` and override `getKey()` so that it applies `sha1()` to it as I frequently get cache keys with colons and other garbage in them. So as with a non-empty prefix, the key gets mutated in this instance, and then gets mutated yet again by the second `getKey()` call, so the lookup key is actually completely different.